Canbiando id a serial

+++ FISEI.ServiceDesk.Domain/Entities/Escalacion.cs
@@ -1,12 +1,10 @@
-using System;
-
 namespace FISEI.ServiceDesk.Domain.Entities;
 
 public class Escalacion
 {
-    public long Id { get; set; }
-    public long IncidenciaId { get; set; }
-    public int Nivel { get; set; } 
+    public int Id { get; set; } // IDENTITY
+    public int IncidenciaId { get; set; }
+    public int Nivel { get; set; }
     public string Motivo { get; set; } = default!;
     public DateTime Fecha { get; set; }
 }
