+++ FISEI.ServiceDesk.Domain/Entities/SLA_Incidencia.cs
@@ -1,11 +1,9 @@
-using System;
-
 namespace FISEI.ServiceDesk.Domain.Entities;
 
 public class SLA_Incidencia
 {
-    public long Id { get; set; }
-    public long IncidenciaId { get; set; }
+    public int Id { get; set; } // IDENTITY
+    public int IncidenciaId { get; set; }
     public DateTime FechaLimiteRespuesta { get; set; }
     public DateTime FechaLimiteResolucion { get; set; }
     public bool CumplidoRespuesta { get; set; }
